requirements.txt
fastapi
uvicorn
python-dotenv
httpx
python-telegram-bot
groq
